fix: résoudre toutes les erreurs clippy pedantic du workspace
- cast_possible_wrap/sign_loss/truncation : remplacés par .cast_signed(),
  .cast_unsigned(), .try_into().unwrap_or(MAX) dans fiscal-engine, edge-api
  et sync-client
- push_str(&format!()) → write!() dans archive.rs et archive_engine
- items_after_statements : struct DiscountEntry déplacé au niveau module
- manual_let_else : if-let → let-else dans config_puller
- match_same_arms : bras identiques fusionnés (TvaRate) dans plusieurs fichiers
- unnecessary_wraps : build_payloads retourne Vec<> directement
- too_many_lines : #[allow] sur run_sync_cycle et create_order_handler
- Suppression des rebindages intermédiaires inutiles dans ms_to_iso8601
- #[allow(dead_code)] sur PayOrderRequest::payment_method (champ API conservé)

cargo clippy --workspace -- -D warnings → 0 erreur
cargo test --workspace → 0 échec

+++ sync-client/src/config.rs
@@ -3,9 +3,9 @@
 //! Configuration du `sync-client` chargée depuis les variables d'environnement.
 //!
 //! ## Variables obligatoires
-//! - `DATABASE_URL` — chemin SQLite de la base du restaurant
+//! - `DATABASE_URL` — chemin `SQLite` de la base du restaurant
 //! - `SUPABASE_URL` — URL de l'instance Supabase
-//! - `SUPABASE_SERVICE_KEY` — clé service_role (jamais la anon key)
+//! - `SUPABASE_SERVICE_KEY` — clé `service_role` (jamais la anon key)
 //!
 //! ## Variables optionnelles (avec défauts)
 //! - `SYNC_INTERVAL_SECONDS` — intervalle entre deux cycles (défaut : 30)
@@ -20,13 +20,13 @@ use crate::error::SyncError;
 /// Configuration complète du sync-client.
 #[derive(Debug, Clone)]
 pub struct SyncConfig {
-    /// URL SQLite locale (ex: `sqlite:./restaurant.db`).
+    /// URL `SQLite` locale (ex: `sqlite:./restaurant.db`).
     pub database_url: String,
 
     /// URL de base Supabase (ex: `https://xxxx.supabase.co`).
     pub supabase_url: String,
 
-    /// Clé service_role Supabase. Ne jamais utiliser la anon key.
+    /// Clé `service_role` Supabase. Ne jamais utiliser la anon key.
     pub supabase_service_key: String,
 
     /// Identifiant unique du restaurant (SIRET ou ID réseau interne).
@@ -156,10 +156,10 @@ fn parse_env_u32(key: &str, default: u32) -> Result<u32, SyncError> {
 /// Jitter pseudo-aléatoire (0..1000ms) basé sur le timestamp système.
 /// Pas de dépendance `rand` pour rester minimal.
 fn jitter_ms() -> u64 {
-    std::time::SystemTime::now()
+    u64::from(std::time::SystemTime::now()
         .duration_since(std::time::UNIX_EPOCH)
         .unwrap_or_default()
-        .subsec_millis() as u64
+        .subsec_millis())
         % 1_000
 }
